requirements.txt
Flask==2.0.1
cryptography==3.4.7
pycryptodome==3.10.1
requests==2.26.0
python-dotenv==0.19.0
pytest==6.2.5
psycopg2-binary==2.9.1 